+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * When closing the individual job log, return to the DI Landing page or View all properly. (UIDATIMP-1156)
 * Replace react-highlighter with a react-17/18 compatible alternative (UIDATIMP-1152)
 * Resizable Panes - Persistence | Use PersistedPaneset smart component for Data import/View all page (UIDATIMP-979)
+* Use API for deleting data import logs on Data Import view all page (UIDATIMP-1127)
 
 ### Bugs fixed:
 * Data Import landing page log shows in old format instead of current format (UIDATIMP-1139)